Quality A/B runs clean; correct a 1.8x data-quality inflation
Two results, one good and one a correction to my own headline figures.

1. First actual quality evidence (router/ab.py, via the local claude CLI so no
   separate credentials are needed):

       model              passed   pass rate   95% CI low     cost
       claude-haiku-4-5    12/12        100%          76%   $0.2191
       claude-opus-5       12/12        100%          76%   $0.9910

   No measured quality gap at 22% of the cost. Scope is comprehension over
   supplied source -- tier T0 -- so it licenses routing read-only lookups to
   Haiku and nothing else. n=12, so the CI is wide; it is a smoke test.
   Actual spend $1.21, against the $0.50 I estimated.

2. 43% of assistant records in the transcripts are DUPLICATES.

   Claude Code writes one JSONL record per content block (text, thinking,
   tool_use), each carrying the same message.usage. Summing raw records
   double-counts: $8,135 raw vs $4,594 deduplicated, a 1.77x inflation.

   Every absolute dollar figure I reported earlier was inflated by that factor.
   The structural claims survive unchanged -- input-side 92%, cache-read 79%,
   output-side 8% -- because proportions are unaffected.

   The important casualty is the output-share claim. Duplicate records inflate
   the output sum but contribute zero context growth (the delta between two
   duplicates is 0), which is exactly how I measured output at ~102% of growth.
   Deduplicated it is ~51%. Tool output -- Bash dominating at 4.5M tokens -- is
   the other half. Terseness therefore caps at about half the pool, and the
   estimate falls from $1,813 to $652.

   A concurrently-running session found this independently; the two measurements
   disagreeing (50% vs 102%) is what surfaced it.

3. One validate claim was measuring the wrong statistic.

   "Sessions long enough for debt to matter" used a plain median and failed at
   80 turns. But sessions under 50 turns are ~47% of sessions and 0.6% of spend,
   while sessions >=200 turns carry 96%. Now cost-weighted: 536 turns, passes.
   The statistic was wrong, not the claim.

Corrected totals: $4,841 measured, $3,558 addressable pool (73%).

+++ router/debt.py
@@ -64,7 +64,7 @@ def token_lifetime_cost(n_tokens: int, remaining_turns: int,
 def breakeven_remaining_turns(model: str = "claude-opus-5", on: date | None = None) -> int:
     """Turns after which re-read cost exceeds generation cost. On Opus 5: 50."""
     r = rate(model, on)
-    return int(round(r.out / (r.inp * CACHE_READ_MULT)))
+    return round(r.out / (r.inp * CACHE_READ_MULT))
 
 
 @dataclass
